# Conflicts:
#	.gitlab-ci.yml
#	common/datatypes/Texture.h
#	common/vc_defines.h

+++ testing/common/TextureTest.cpp
@@ -66,7 +66,7 @@ struct EmptyConstructorTextureFixture {
 
     volcart::Texture _Texture;
     std::string _ID, _Type, _TruncatedID;
-    int _NumberOfImages;
+    size_t _NumberOfImages;
 };
 
 /*
